Improve code organization (#185)
