+++ configs/entity-gym/minefield.ron
@@ -2,19 +2,21 @@
 ExperimentConfig(
     env: (
         id: "Minefield",
-        num_envs: 512,
-        num_steps: 512,
-        processes: 16,
         kwargs: "{\"max_mines\": 10}"
+    ),
+    rollout: (
+        num_envs: 512,
+        steps: 512,
+        processes: 16,
     ),
     total_timesteps: 10000000,
     net: (
         d_model: 32,
-        n_layer: 2
-    ),
-    translation: (
-        reference_entity: "Vehicle",
-        position_features: ["x_pos", "y_pos"],
+        n_layer: 2,
+        translation: (
+            reference_entity: "Vehicle",
+            position_features: ["x_pos", "y_pos"],
+        ),
     ),
     optim: (
         bs: 1024,
